// src/lib/api/generated/models/computedFieldDefinitionRequest.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Hubuum REST API
 * OpenAPI documentation for the Hubuum REST service.
 * OpenAPI spec version: 0.0.2
 */
import type { ComputedFieldDefinitionRequestOperation } from './computedFieldDefinitionRequestOperation';
import type { ComputedResultType } from './computedResultType';

export interface ComputedFieldDefinitionRequest {
  description?: string;
  enabled?: boolean;
  key: string;
  label: string;
  operation: ComputedFieldDefinitionRequestOperation;
  result_type: ComputedResultType;
}