// src/lib/api/generated/models/computedResultType.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Hubuum REST API
 * OpenAPI documentation for the Hubuum REST service.
 * OpenAPI spec version: 0.0.2
 */

export type ComputedResultType = typeof ComputedResultType[keyof typeof ComputedResultType];


export const ComputedResultType = {
  string: 'string',
  number: 'number',
  integer: 'integer',
  boolean: 'boolean',
  object: 'object',
  array: 'array',
} as const;